requirements.txt
Flask==2.1.2
google-cloud-firestore==2.5.3
requests==2.26.0
opencv-python==4.5.3.56
face_recognition==1.3.0
